Resolve problems with social_access callback.
This bug presented in a strange manner by occurring at times
when it was not apparent that the social_access callback
should have even been called. Performing traceback revealed
that the admin_menu was calling the social admin menu which
had errors in the way it was passing arguments to
social_access. social_access also had errors in how it was
receiving parameters in that it was always expecting a social
object even when social_type string names were passed.